# services/api/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "exam-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v2/auth/register/email": {
      "post": {
        "tags": [
          "identity-v2"
        ],
        "summary": "Register Email",
        "operationId": "register_email_api_v2_auth_register_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEmailRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/register/phone": {
      "post": {
        "tags": [
          "identity-v2"
        ],
        "summary": "Register Phone",
        "operationId": "register_phone_api_v2_auth_register_phone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPhoneRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/login": {
      "post": {
        "tags": [
          "identity-v2"
        ],
        "summary": "Login",
        "operationId": "login_api_v2_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/logout": {
      "post": {
        "tags": [
          "identity-v2"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v2_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthAckV2"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/auth/session": {
      "get": {
        "tags": [
          "identity-v2"
        ],
        "summary": "Get Session",
        "operationId": "get_session_api_v2_auth_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionStateResponseV2"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/auth/send-code": {
      "post": {
        "tags": [
          "identity-v2"
        ],
        "summary": "Send Code",
        "operationId": "send_code_api_v2_auth_send_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCodeRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendCodeResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/verify-code": {
      "post": {
        "tags": [
          "identity-v2"
        ],
        "summary": "Verify Code",
        "operationId": "verify_code_api_v2_auth_verify_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyCodeResponseV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/auth/reset-password": {
      "post": {
        "tags": [
          "identity-v2"
        ],
        "summary": "Reset Password",
        "operationId": "reset_password_api_v2_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthAckV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/notes": {
      "get": {
        "tags": [
          "notes-v2-skeleton"
        ],
        "summary": "List Notes",
        "operationId": "list_notes_api_v2_notes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteListResponseV2"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "notes-v2-skeleton"
        ],
        "summary": "Create Note",
        "operationId": "create_note_api_v2_notes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreateRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDetailV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/notes/{note_id}": {
      "get": {
        "tags": [
          "notes-v2-skeleton"
        ],
        "summary": "Get Note",
        "operationId": "get_note_api_v2_notes__note_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDetailV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "notes-v2-skeleton"
        ],
        "summary": "Update Note",
        "operationId": "update_note_api_v2_notes__note_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Note Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteUpdateRequestV2"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDetailV2"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthAckV2": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "title": "AuthAckV2"
      },
      "AuthSessionResponseV2": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserV2"
          },
          "session": {
            "$ref": "#/components/schemas/AuthSessionV2"
          },
          "csrfToken": {
            "type": "string",
            "title": "Csrftoken"
          }
        },
        "type": "object",
        "required": [
          "user",
          "session",
          "csrfToken"
        ],
        "title": "AuthSessionResponseV2"
      },
      "AuthSessionStateResponseV2": {
        "properties": {
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserV2"
          },
          "session": {
            "$ref": "#/components/schemas/AuthSessionV2"
          }
        },
        "type": "object",
        "required": [
          "authenticated",
          "user",
          "session"
        ],
        "title": "AuthSessionStateResponseV2"
      },
      "AuthSessionV2": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "issuedAt": {
            "type": "string",
            "title": "Issuedat"
          },
          "expiresAt": {
            "type": "string",
            "title": "Expiresat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "token",
          "issuedAt",
          "expiresAt"
        ],
        "title": "AuthSessionV2"
      },
      "AuthUserV2": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "publicId": {
            "type": "string",
            "title": "Publicid"
          },
          "displayName": {
            "type": "string",
            "title": "Displayname"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "publicId",
          "displayName",
          "isActive",
          "createdAt"
        ],
        "title": "AuthUserV2"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequestV2": {
        "properties": {
          "identifier": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Identifier"
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "password"
        ],
        "title": "LoginRequestV2"
      },
      "NoteCreateRequestV2": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "NoteCreateRequestV2"
      },
      "NoteDetailV2": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "body",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "NoteDetailV2"
      },
...
      "NoteListResponseV2": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NoteItemV2"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "title": "NoteListResponseV2"
      },
      "NoteUpdateRequestV2": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "NoteUpdateRequestV2"
      },
...
      "RegisterEmailRequestV2": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 6,
            "title": "Password"
          },
          "displayName": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Displayname",
            "default": "New User"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterEmailRequestV2"
      },
      "RegisterPhoneRequestV2": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 6,
            "title": "Phone"
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 6,
            "title": "Password"
          },
          "displayName": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Displayname",
            "default": "New User"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "password"
        ],
        "title": "RegisterPhoneRequestV2"
      },
      "ResetPasswordRequestV2": {
        "properties": {
          "identifier": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Identifier"
          },
          "code": {
            "type": "string",
            "maxLength": 16,
            "minLength": 4,
            "title": "Code"
          },
          "newPassword": {
            "type": "string",
            "maxLength": 255,
            "minLength": 6,
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "code",
          "newPassword"
        ],
        "title": "ResetPasswordRequestV2"
      },
...
      "SendCodeRequestV2": {
        "properties": {
          "targetKind": {
            "type": "string",
            "enum": [
              "email",
              "phone"
            ],
            "title": "Targetkind"
          },
          "targetValue": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Targetvalue"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "register",
              "reset_password",
              "login",
              "bind"
            ],
            "title": "Purpose"
          }
        },
        "type": "object",
        "required": [
          "targetKind",
          "targetValue",
          "purpose"
        ],
        "title": "SendCodeRequestV2"
      },
      "SendCodeResponseV2": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "delivery": {
            "type": "string",
            "title": "Delivery"
          },
          "devCode": {
            "type": "string",
            "title": "Devcode"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "purpose",
          "delivery",
          "devCode"
        ],
        "title": "SendCodeResponseV2"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyCodeRequestV2": {
        "properties": {
          "targetKind": {
            "type": "string",
            "enum": [
              "email",
              "phone"
            ],
            "title": "Targetkind"
          },
          "targetValue": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Targetvalue"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "register",
              "reset_password",
              "login",
              "bind"
            ],
            "title": "Purpose"
          },
          "code": {
            "type": "string",
            "maxLength": 16,
            "minLength": 4,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "targetKind",
          "targetValue",
          "purpose",
          "code"
        ],
        "title": "VerifyCodeRequestV2"
      },
      "VerifyCodeResponseV2": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "verified",
          "message"
        ],
        "title": "VerifyCodeResponseV2"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}